[MISC] Use same type for kmer_counts.

+++ test/api/layout/execute_with_estimation_test.cpp
@@ -66,7 +66,7 @@ TEST(execute_estimation_test, many_ubs)
     std::filesystem::path const stats_file{layout_file.get_path().string() + ".stats"};
 
     std::vector<std::string> many_filenames;
-    std::vector<uint64_t> many_kmer_counts;
+    std::vector<size_t> many_kmer_counts;
 
     // There are 20 files with a count of {100,200,300,400} each. There are 16 files with count 500.
     for (size_t i{0}; i < 96u; ++i)
